Overhauling neuron state updates to use current time in timesteps. Major individual changes. Fix for generator neuron code w/ synaptic update.

+++ Spike/Backend/CUDA/Neurons/SpikingNeurons.hpp
@@ -61,7 +61,7 @@ namespace Backend {
        */
       void copy_constants_to_device(); // Not virtual
 
-      void state_update(float current_time_in_seconds, float timestep) override;
+      void state_update(int current_time_in_seconds, float timestep) override;
       
     };
